requirements.txt
flask
flask_sqlalchemy
flask_bcrypt
flask_login
Pillow
flask_wtf
email_validator